feat: add more destructuring
