+++ src/App.js
@@ -17,8 +17,8 @@ const App = (props) => (
       <Header />
       <Navbar />
       <div className="app-wrapper-content">
-        <Route path="/profile" render={() => <Profile posts={props.state.profilePage.posts} />} />
-        <Route path="/dialogs" render={() => <Dialogs dialogs={props.state.dialogsPage.dialogs} messages={props.state.dialogsPage.messages} />} />
+        <Route path="/profile" render={() => <Profile state={props.state.profilePage} />} />
+        <Route path="/dialogs" render={() => <Dialogs state={props.state.dialogsPage} />} />
         <Route path="/news" render={() => <News />} />
         <Route path="/music" render={() => <Music />} />
         <Route path="/settings" render={() => <Settings />} />
