+++ src/components/Dialogs/Message/Message.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+import style from '../Dialogs.module.css';
+
+const Message = (props) => <div className={style.message}>{props.message}</div>;
+
+export default Message;
