+++ src/components/Profile/Profile.js
@@ -6,7 +6,7 @@ import ProfileInfo from './ProfileInfo/ProfileInfo';
 const Profile = (props) => (
   <div>
     <ProfileInfo />
-    <MyPosts posts={props.posts} />
+    <MyPosts posts={props.state.posts} />
   </div>
 );
